+++ public/css/index.css
@@ -1,6 +1,3 @@
-@import url(/public/icons/search.css);
-@import url(/public/css/background-image.css);
-
 .index--banner {
 	height: 400px;
 }
