+++ src/Controller/Page/Logout.php
@@ -1,15 +1,12 @@
 <?php namespace Controller\Page;
 
+use Modules\Account\Session;
 use Modules\Kernel\Page;
+use Modules\Kernel\Response;
 
 class Logout extends Page {
 	public array $publicaciones;
 	function __construct() {
-		unset($_SESSION['uid']);
-		session_start();
-		session_unset();
-		session_write_close();
-		session_destroy();
-		header('Location: /');
+		Session::close();
 	}
 }
