+++ src/Modules/Kernel/Storage.php
@@ -14,6 +14,11 @@ abstract class Storage {
 	static function connected(): bool {
 		return isset(self::$driver);
 	}
+
+	static function stop() {
+		if(self::connected())
+			self::$driver->close();
+	}
 }
 
 interface Query {
