--- src/Modules/Response.php
@@ -1,6 +1,0 @@
-<?php namespace Modules\Kernel;
-
-class Response {
-	static function send() {
-	}
-}
