+++ src/Templates/create_publication.phtml
@@ -1,0 +1,16 @@
+<div class="container">
+	<h1>Crear una publicación</h1>
+
+	<form enctype="multipart/form-data" manager="form" action="/admin/publicaciones">
+		<label>Titulo: </label>
+		<input type="text" name="title">
+		<label>Imagen: </label>
+		<input type="file" name="image">
+		<label>Cuerpo: </label>
+		<textarea name="body" rows="12"></textarea>
+		<div>
+			<button type="submit primary">Crear</button>
+			<a class="button" href="..">Cancelar</a>
+		</div>
+	</form>
+</div>
